+++ sync-for-payroll/RELEASES.md
@@ -79,3 +79,13 @@ Based on:
 - [typescript v1.4.0] sync-for-payroll
 ### Releases
 - [NPM v1.4.0] https://www.npmjs.com/package/@codat/sync-for-payroll/v/1.4.0 - sync-for-payroll
+
+## 2023-10-03 00:05:29
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Payroll.yaml
+- Speakeasy CLI 1.93.1 (2.144.7) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.4.1] sync-for-payroll
+### Releases
+- [NPM v1.4.1] https://www.npmjs.com/package/@codat/sync-for-payroll/v/1.4.1 - sync-for-payroll